derectifying usemusicplayer into playeractions
