+++ src/app/_components/ConfirmationDialog.tsx
@@ -21,8 +21,8 @@ export const ConfirmationDialog = ({
   return (
     <Dialog open={open} onOpenChange={onOpenChange}>
       <DialogContent>
-        <DialogHeader>{text}</DialogHeader>
-        <DialogFooter>
+        <DialogHeader className="sm:text-center">{text}</DialogHeader>
+        <DialogFooter className="sm:justify-center">
           <DialogClose asChild>
             <Button variant="outline">Cancel</Button>
           </DialogClose>
